+++ doc/news.rst
@@ -23,6 +23,12 @@ In summary, you should always be able to upgrade to a release which only
 increments ``patch``, but if either of the other two components have changed,
 you should review this document and see if your consfig needs updating.
 
+1.1.0 (unreleased)
+------------------
+
+- API change: DISK:WITH-OPENED-VOLUMES now includes volumes that were already
+  open, and their parents, in the connattrs.
+
 1.0.3 (2022-06-29)
 ------------------